+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 dev:
   - generate error message if "validator credentials set" process fails to generate any credentials
   - allow import of accounts with null name field in their keystore
+  - show text of execution payload extra data if available
 
 1.28.0:
   - support additional mnemonic word list languages